+++ verify/obj_test.go
@@ -53,42 +53,6 @@ func TestObj(t *testing.T) {
 		})
 	})
 
-	t.Run("Zero", func(t *testing.T) {
-		t.Run("Passed", func(t *testing.T) {
-			got := A{}
-			msg := verify.Obj(got).Zero()
-			assertPassed(t, msg)
-		})
-		t.Run("Failed", func(t *testing.T) {
-			got := A{Str: "wrong"}
-			msg := verify.Obj(got).Zero()
-			assertFailed(t, msg, "not a zero value (-want +got):\n")
-		})
-		t.Run("nil", func(t *testing.T) {
-			var got *A
-			msg := verify.Obj(got).Zero()
-			assertPassed(t, msg)
-		})
-	})
-
-	t.Run("NonZero", func(t *testing.T) {
-		t.Run("Passed", func(t *testing.T) {
-			got := A{Str: "string"}
-			msg := verify.Obj(got).NonZero()
-			assertPassed(t, msg)
-		})
-		t.Run("Failed", func(t *testing.T) {
-			got := A{}
-			msg := verify.Obj(got).NonZero()
-			assertFailed(t, msg, "a zero value")
-		})
-		t.Run("nil", func(t *testing.T) {
-			var got *A
-			msg := verify.Obj(got).NonZero()
-			assertFailed(t, msg, "a zero value")
-		})
-	})
-
 	t.Run("Check", func(t *testing.T) {
 		t.Run("Passed", func(t *testing.T) {
 			fn := func(x A) verify.FailureMessage {